Add # of PRN map entries (#20)
First part of
https://github.com/ParkingReformNetwork/organization-dashboard/issues/15.
(Still missing a script to pull in historical data one time.)

--

Most importantly, this sets up the architecture of the project. The
program is run once, rather than continually maintaining a server. Each
time you run it, you specify which services you want to run, like
`instagram` and `map-projects`.

That means that the program caller is responsible for coordinating the
timing and frequency of each service, such as GitHub Actions setting
some up to run every day and others once a week. The program is
oblivious to that, which is a cleaner architecture.

Each service has a dedicated file that does whatever it needs to get its
`Point`s to pass to InfluxDB. `index.js` knows when to call into these
helper functions, then combines all the results and writes them to
InfluxDB. Good separation of concerns.

The CLI interface looks like `npm start -- --services map-projects
instagram`. I went with this `--services` arg, rather than e.g.
`--github --instagram`, because it allows us to add follow up options
like `--dry-run`.