Test map-projects service (#31)
Closes
https://github.com/ParkingReformNetwork/organization-dashboard/issues/25.

These tests are useful because we'll want to reuse `parsePoints` when
getting the `--historical` data for the service.

## Uses Playwright

I went with Playwright as the test runner, rather than Jest, for
consistency with the rest of the Parking Reform projects.

## Uses new namespace scheme

See
https://github.com/ParkingReformNetwork/organization-dashboard/issues/29
for the rationale.

## Uses real data

I pulled actual data from the map projects, rather than creating my own
fake data. I figured this is useful to capture the complexity of the
real world.