Adopt TypeScript (#33)
Closes
https://github.com/ParkingReformNetwork/organization-dashboard/issues/26.

I decided TypeScript is worth it because it makes code more
self-documenting and the IDE experience so much better. It also catches
a class of bugs.

While their is some learning curve, all our contributors have used typed
languages like Swift and C++, so the curve is less intense and types may
even be more familiar.

I didn't decide to adopt it for mandates map & parking lots map because
they have substantially less JavaScript code. But here, this is a pure
JavaScript project and so I think worth it. (And it may be worth
revisiting those projects since Playwright and Parcel support TypeScript
so nicely.)