Add `runProcess` util (#37)
Node.js's interface for running subprocesses is not very ergonomic. This
wraps it in a helper function that allows us to use async/await, which
is much more natural.

It intentionally errors on failures to keep things simple. We don't need
to handle failed subprocesses in this repository.

It returns `[stdout, stderr]` so that we can e.g. parse the result of
`git log`.